Move UserDirectory.java into new directory repository.
