+++ src/main/java/com/coworking_service/Controller.java
@@ -5,7 +5,7 @@ import com.coworking_service.in.UserInputHandler;
 import com.coworking_service.repository.BookingDirectory;
 import com.coworking_service.model.CoworkingSpace;
 import com.coworking_service.model.User;
-import com.coworking_service.model.UserDirectory;
+import com.coworking_service.repository.UserDirectory;
 import com.coworking_service.model.enums.MessageType;
 import com.coworking_service.model.enums.Role;
 import com.coworking_service.out.UserOutputHandler;
